requirements.txt
py-cord[voice]
deepgram-sdk
python-dotenv
watchdog

